backend/requirements.txt
Flask==2.3.3
cryptography==41.0.4
flask-cors==4.0.0
PyJWT==2.8.0